add `blog_post` to adm

+++ StudyApp/admin/routes.py
@@ -5,24 +5,29 @@ from StudyApp.admin.forms import *
 
 admin = Blueprint("admin", __name__)
 
+def verify_admin():
+    if not current_user.is_authenticated or not current_user.is_admin:
+        abort(403)
+    else:
+        return True
+
 @admin.route("/admin/")
 def admin_panel():
     Stats.query.filter_by(user_id=2).delete()
-    if not current_user.is_authenticated or not current_user.is_admin:
-        abort(403)
+    verify_admin()
     return render_template("admin/index.html")
 
 @admin.route("/admin/users/")
 def users():
-    if not current_user.is_authenticated or not current_user.is_admin:
-        abort(403)
+    verify_admin()
     users = User.query.all()
     stats = Stats.query.all()
     users_stats = zip(users, stats)
-    return render_template("admin/users.html", users_stats=users_stats)
+    return render_template("admin/users/users.html", users_stats=users_stats)
 
 @admin.route("/admin/users/delete/<int:user_id>", methods=["GET", "POST"])
 def users_delete(user_id):
+    verify_admin()
     username = User.query.filter_by(id=user_id).first().username
     User.query.filter_by(id=user_id).delete()
     Stats.query.filter_by(user_id=user_id).delete()
@@ -32,6 +37,7 @@ def users_delete(user_id):
 
 @admin.route("/admin/users/edit/<int:user_id>", methods=["GET", "POST"])
 def users_edit(user_id):
+    verify_admin()
     user = User.query.filter_by(id=user_id).first_or_404()
     form = UpdateUserForm(user=user)
     if form.validate_on_submit():
@@ -46,10 +52,11 @@ def users_edit(user_id):
         form.email.data = user.email
         form.username.data = user.username
         form.is_admin.data = user.is_admin
-    return render_template("admin/users_edit.html", form=form)
+    return render_template("admin/users/users_edit.html", form=form)
 
 @admin.route("/admin/users/add/", methods=["GET", "POST"])
 def users_add():
+    verify_admin()
     form = NewUserForm()
     if form.validate_on_submit():
         hashed_password = bcrypt.generate_password_hash(form.password.data).decode("utf-8")
@@ -62,19 +69,19 @@ def users_add():
         db.session.commit()
         flash({"title": "Congratulations!", "message": f"Account created for {form.username.data}!"}, "success")
         return redirect(url_for("admin.users"))
-    return render_template("admin/users_add.html", form=form)
+    return render_template("admin/users/users_add.html", form=form)
 
 # CATEGORIES
 
 @admin.route("/admin/category/")
 def categories():
-    if not current_user.is_authenticated or not current_user.is_admin:
-        abort(403)
+    verify_admin()
     categories = Category.query.all()
-    return render_template("admin/categories.html", categories=categories)
+    return render_template("admin/categories/categories.html", categories=categories)
 
 @admin.route("/admin/categories/delete/<int:category_id>", methods=["GET", "POST"])
 def categories_delete(category_id):
+    verify_admin()
     name = Category.query.filter_by(id=category_id).first().name
     Category.query.filter_by(id=category_id).delete()
     db.session.commit()
@@ -83,6 +90,8 @@ def categories_delete(category_id):
 
 @admin.route("/admin/categories/edit/<int:category_id>", methods=["GET", "POST"])
 def categories_edit(category_id):
+    verify_admin()
+
     category = Category.query.get_or_404(category_id)
     form = UpdateCategoryForm(category=category)
 
@@ -94,10 +103,11 @@ def categories_edit(category_id):
     else:
         form.name.data = category.name
 
-    return render_template("admin/categories_edit.html", form=form)
+    return render_template("admin/categories/categories_edit.html", form=form)
 
 @admin.route("/admin/categories/add/", methods=["GET", "POST"])
 def categories_add():
+    verify_admin()
     form = NewCategoryForm()
     if form.validate_on_submit():
         category = Category(name=form.name.data)
@@ -105,4 +115,39 @@ def categories_add():
         db.session.commit()
         flash({"title": "Congratulations!", "message": f"{form.name.data} created!"}, "success")
         return redirect(url_for("admin.categories"))
-    return render_template("admin/categories_add.html", form=form)
+    return render_template("admin/categories/categories_add.html", form=form)
+
+# BLOG POSTS
+
+@admin.route("/admin/blog_post/")
+def blog_post():
+    verify_admin()
+    blog_post = BlogPost.query.all()
+    return render_template("admin/blog_post/blog_post.html", blog_post=blog_post, Category=Category)
+
+@admin.route("/admin/blog_post/delete/<int:blog_post_id>", methods=["GET", "POST"])
+def blog_post_delete(blog_post_id):
+    verify_admin()
+    name = BlogPost.query.filter_by(id=blog_post_id).first().title
+    BlogPost.query.filter_by(id=blog_post_id).delete()
+    db.session.commit()
+    flash({"title": "Congratulations!", "message": f"{name} deleted!"}, "success")
+    return redirect(url_for("admin.blog_post"))
+
+"""@admin.route("/admin/blog_post/edit/<int:category_id>", methods=["GET", "POST"])
+def blog_post_edit(category_id):
+    verify_admin()
+
+    category = Category.query.get_or_404(category_id)
+    form = UpdateCategoryForm(category=category)
+
+    if form.validate_on_submit():
+        category.name = form.name.data
+        db.session.commit()
+        flash({"title": "Congratulations!", "message": f"{form.name.data} edited!"}, "success")
+        return redirect(url_for("admin.categories"))
+    else:
+        form.name.data = category.name
+
+    return render_template("admin/categories/categories_edit.html", form=form)
+"""
